1 .除了报销生成 界面 所有总额统计 修改为CNY 金额 2. 优化记一单 保存 合法性检查逻辑，总额不允许为零 3. 对部分UI界面代码 进行了封装。

+++ HANDMobileExp/Frameworks/LittleMvc/LMCellStypeItem.h
@@ -13,7 +13,7 @@
 
 @property (nonatomic, assign) id        delegate;
 @property (nonatomic, assign) SEL       selector;
-@property (nonatomic) float  amount;
+@property (nonatomic) double  amount;
 @property (nonatomic,strong) NSString * expense_type_desc;
 @property (nonatomic,strong) NSString * line_desc;
 @property (nonatomic,strong) NSNumber *  primary_id;
